adapt to changes in `gix-date`

+++ cargo-smart-release/src/utils.rs
@@ -300,7 +300,7 @@ mod tests {
 }
 
 pub fn time_to_offset_date_time(time: gix::date::Time) -> OffsetDateTime {
-    time::OffsetDateTime::from_unix_timestamp(time.seconds as i64)
+    time::OffsetDateTime::from_unix_timestamp(time.seconds)
         .expect("always valid unix time")
         .replace_offset(time::UtcOffset::from_whole_seconds(time.offset).expect("valid offset"))
 }
